requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
APScheduler==3.10.4
pydantic-settings==2.2.1
PyYAML==6.0.1
redis==5.0.1
starlette==0.37.2
